feat: added Item entity and constructors

+++ src/main/java/com/autonexo/inventory/domain/model/commands/CreateItemInInventoryCommand.java
@@ -1,6 +1,6 @@
 package com.autonexo.inventory.domain.model.commands;
 
-public record CreateItemInInventoryCommand(String name, String description, Long quantity, Long inventoryId) {
+public record CreateItemInInventoryCommand(String name, String description, Integer quantity, Long inventoryId) {
     public CreateItemInInventoryCommand {
         if (inventoryId == null || inventoryId <= 0) {
             throw new IllegalArgumentException("inventoryId cannot be null or less than 1");
